+++ setup.sh
@@ -444,6 +444,10 @@ function wizard() {
   read -r -p "$(echo -e "Install with ${BUNLINE}sql(cl)${NC} or ${BUNLINE}sqlplus${NC}? [${BGRAY}${local_sqlcli}${NC}]: ")" wiz_sqlcli
   wiz_sqlcli=${wiz_sqlcli:-"${local_sqlcli}"}
 
+  local local_logpath=${LOG_PATH-""}
+  read -r -p "$(echo -e "Enter path to place logfiles into after installation.? [${BGRAY}${local_logpath}${NC}]: ")" wiz_logpath
+  wiz_logpath=${wiz_logpath:-"${local_logpath}"}
+
   if [[ ${apply_only} == "NO" ]]; then
 
     if [[ ${env_only} == "NO" ]]; then
@@ -470,7 +474,8 @@ function write_apply() {
      [[ -z ${wiz_db_admin_user+x} ]] || \
      [[ -z ${wiz_depot_path+x} ]] ||\
      [[ -z ${wiz_stage+x} ]] ||\
-     [[ -z ${wiz_sqlcli+x} ]]
+     [[ -z ${wiz_sqlcli+x} ]] ||\
+     [[ -z ${wiz_logpath+x} ]]
     then
     echo_error "Not all vars set"
     exit 1
@@ -515,6 +520,10 @@ function write_apply() {
     echo ""
     echo "# TEAMS Channel to Post to on success"
     echo "TEAMS_WEBHOOK_URL="
+    echo ""
+    echo "# Path to copy logs to after installation"
+    echo "LOG_PATH=${wiz_logpath}"
+
   } > apply.env
 }