+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - New: Create readme.md when generating project
 - New: apply.sh can now run a stepwise installation using flag --stepwise
 - New: setup.sh --apply will only generate apply.env file
+- New: added new env var LOG_PATH which defines an additional path to write the logs to
 - Fix: Synonyms will be dropped in init mode as well
 
 ## [2.2.0 - 2023-07-01]